% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{check_missingness}
\alias{check_missingness}
\title{check_missingness}
\usage{
check_missingness(dat, cols, out = "drop", prop_good = 0.5, llod = 1e+06)
}
\arguments{
\item{dat}{data.frame or matrix to check}

\item{cols}{Vector of columns to check for missingness}

\item{out}{Character value to label output column}

\item{prop_good}{Numeric proportion of good values to keep. Default, 0.5, indicates at least 50 percent of values must be non-missing and above the lower limit of detection.}

\item{llod}{Numeric lower limit of detection. Any value below this will be considered missing.}
}
\value{
data frame or matrix with missingness information added as a new column
}
\description{
Check for missingness in data for a group of columns and flag rows with too many missing or low values.
}
\examples{
set.seed(294873)
dat <- data.frame(pep = c(1, 2, 3),
                  A = runif(3) * 1e7,

                  # middle value is missing and last value is too small
                  B = runif(3) * c(1e7, NA, 1e4),

                  # middle value is missing
                  C = runif(3) * c(1e7, NA, 1e7),

                  # last value is too small
                  D = runif(3) * c(1e7, 1e7, 1e4))

# all rows have at least two good values, so no rows are flagged to drop
check_missingness(dat, 2:5)

# the second and third rows have less than 60\% good values, so they are flagged to drop
check_missingness(dat, 2:5, prop_good = 0.6)

}
