% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_progress.R
\name{check_progress}
\alias{check_progress}
\title{check_progress
Check progress of previous runs and determine what needs to be done}
\usage{
check_progress(config)
}
\arguments{
\item{config}{list of configuration options}
}
\value{
data.frame with progress information
}
\description{
check_progress
Check progress of previous runs and determine what needs to be done
}
