% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{fld_to_pld}
\alias{fld_to_pld}
\title{fld_to_pld
Convert FeatureLevelData to ProteinLevelData}
\usage{
fld_to_pld(fld)
}
\arguments{
\item{fld}{FeatureLevelData}
}
\value{
ProteinLevelData data.frame
}
\description{
fld_to_pld
Convert FeatureLevelData to ProteinLevelData
}
