% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_MSfmtR}
\alias{load_MSfmtR}
\title{load_MSfmtR}
\usage{
load_MSfmtR(lib.loc, commit)
}
\arguments{
\item{lib.loc}{Character, specifying the location of a local library to use for installation (i.e. to avoid overwriting the system-wide installation)}

\item{commit}{Character, specifying the GitHub commit to use}
}
\value{
Logical value indicating the correct version of the package was successfully loaded
}
\description{
Load a specific version of MSfmtR
}
