% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_samples_groups.R
\name{map_samples}
\alias{map_samples}
\alias{map_groups}
\title{map_samples
Map originalRUN IDs to sample IDs}
\usage{
map_samples(data, config)

map_groups(data, config)
}
\arguments{
\item{data}{A data.frame from an MSstats object - either `MSstats_obj$FeatureLevelData` or `MSstats_obj$ProteinLevelData`}

\item{config}{A list containing the configuration parameters}
}
\value{
An update to `data` with modified sample IDs. For sample mapping, a new variable, `sample` is added, and for group mapping a new variable `group` is added (not overwritting the variable `GROUP` that should be included in `data`).
}
\description{
map_samples
Map originalRUN IDs to sample IDs

map_groups
Map originalRUN IDs to group IDs
}
