% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_proteins.R
\name{process_proteins}
\alias{process_proteins}
\title{process_proteins
Process proteins from raw data}
\usage{
process_proteins(
  data,
  peptides,
  config,
  stage = file.path(config$output_dir, config$protein_checkpoint),
  save_intermediate = TRUE
)
}
\arguments{
\item{data}{MSstats formatted data}

\item{peptides}{processed and formatted peptides data}

\item{config}{list of configuration parameters}

\item{stage}{character path to checkpoint file}

\item{save_intermediate}{logical save intermediate data}
}
\value{
data frame of processed proteins data
}
\description{
process_proteins
Process proteins from raw data
}
\details{
This function processes MSstats formatted data and returns proteins ready for ProtResDash. If save_intermediate is TRUE, the processed data are also saved to the checkpoint file.
}
