% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{raw_to_fld}
\alias{raw_to_fld}
\title{raw_to_fld
Convert raw data to FeatureLevelData}
\usage{
raw_to_fld(raw, format = "MSstats", config = list(lloq = 0, uloq = Inf))
}
\arguments{
\item{raw}{data.frame}

\item{format}{character describing the format of the raw data}

\item{config}{list of configuration parameters}
}
\value{
FeatureLevelData data.frame
}
\description{
raw_to_fld
Convert raw data to FeatureLevelData
}
\details{
The only config values used are `lloq` and `uloq`. These are used to filter out-of-spec data.

Currently supported formats are
'MSstats', the MSstats report exported by Spectronaut, and
'other', the ... report exported by Spectronaut.
}
\references{
See documentation of `MSstats::dataProcess` at https://www.bioconductor.org/packages/devel/bioc/vignettes/MSstats/inst/doc/MSstats.html)
}
