% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PD.R
\name{read_PD}
\alias{read_PD}
\title{read_PD}
\usage{
read_PD(config, file = NULL, ...)
}
\arguments{
\item{config}{List containing configuration parameters}

\item{file}{Character, path to input file. If NULL (default), parameters in `config` are used.}

\item{...}{additional agruments to pass to `updt_config`}
}
\value{
A list with two data.frames, `proteins` and `peptides`
}
\description{
Read excel files exported from Proteome Discoverer
}
\details{
The input file should be an excel spreadsheet <...describe formatting...>
Protein and peptide rows are written to a temporary directory and read in separately to maintain correct column types.

 Configuration parameters relevant to this function include: `input_dir`, `in_file`, `sheet`.
 See `?updt_config` for more configuration details.
}
