% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tukey_biweight_mean}
\alias{tukey_biweight_mean}
\title{tukey_biweight_mean
Tukey Biweight Mean}
\source{
Function definition provided by Gemini 2.0 Flash
}
\usage{
tukey_biweight_mean(x, c = 4.685)
}
\arguments{
\item{x}{numeric vector}

\item{c}{tuning constant}
}
\value{
numeric
}
\description{
tukey_biweight_mean
Tukey Biweight Mean
}
