% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_formatR.R
\name{updt_config}
\alias{updt_config}
\title{updt_config
Update configuration settings (primarily used for local changes to defaults)}
\usage{
updt_config(config, ...)
}
\arguments{
\item{config}{list of configuration settings}

\item{...}{named arguments to update}
}
\description{
updt_config
Update configuration settings (primarily used for local changes to defaults)
}
