% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{check_missingness}
\alias{check_missingness}
\alias{check_missingness_chr}
\alias{check_missingness_num}
\title{check_missingness}
\usage{
check_missingness(dat, cols, ...)

check_missingness_chr(dat, cols, prop_good = 1, valid_chr = NULL, ...)

check_missingness_num(dat, cols, prop_good = 1, llod = 0, ...)
}
\arguments{
\item{dat}{data.frame or matrix to check}

\item{cols}{Vector of columns to check for missingness}

\item{...}{Additional arguments passed to `check_missingness_num()` or `check_missingness_chr()`.}

\item{prop_good}{Numeric proportion of good values to keep.}

\item{valid_chr}{Character vector of valid values for character columns. If `NULL` (default), all non-missing values are considered valid.}

\item{llod}{Numeric lower limit of detection. Any value below this will be considered missing.}
}
\value{
vector of logicals indicating which rows to filter / drop. In other words, rows with `TRUE` should be dropped.
}
\description{
Check for missingness in data for a group of columns and flag rows with too many missing or low values.
}
\details{
By default, this function ignores the `llod` argument (since `llod = 0`) and requires all values to be non-missing.
If the lower limit of detection for your assay is 1e6, for example, you can set `llod = 1e6` to treat any value below this as missing.
To allow for missing values, set `prop_good` to something less than 1.
For example, 0.5 indicates at least 50 percent of values on a row must be non-missing and above the lower limit of detection.
}
\examples{
set.seed(294873)
dat <- data.frame(pep = c(1, 2, 3),
                  A = runif(3) * 1e7,

                  # middle value is missing and last value is too small
                  B = runif(3) * c(1e7, NA, 1e4),

                  # middle value is missing
                  C = runif(3) * c(1e7, NA, 1e7),

                  # last value is too small
                  D = runif(3) * c(1e7, 1e7, 1e4))

# all rows have at least two good values, so no rows are flagged to drop
check_missingness(dat, 2:5, prop_good = 0.5, llod = 1e6)

# the second and third rows have less than 60\% good values, so they are flagged to drop
check_missingness(dat, 2:5, prop_good = 0.6, llod = 1e6)

}
