% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_formatR.R
\name{configure_formatR}
\alias{configure_formatR}
\title{configure_formatR
Default configuration for MSfmtR}
\usage{
configure_formatR(config_file = NULL, args = NULL)
}
\arguments{
\item{config_file}{path to a yaml file with configuration settings}

\item{args}{list of arguments to override defaults}
}
\value{
list of configuration settings
}
\description{
configure_formatR
Default configuration for MSfmtR
}
