% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_completeness}
\alias{filter_completeness}
\title{filter_completeness}
\usage{
filter_completeness(
  dat,
  cols,
  filter_cols,
  ratio = FALSE,
  conditions = gsub("drop_", "", names(dat)[filter_cols])
)
}
\arguments{
\item{dat}{data.frame to filter}

\item{cols}{Integer, columns to filter}

\item{filter_cols}{Integer, columns of `dat` with missingness information (see Details).}

\item{ratio}{Logical value indicating whether the columns to be filtered are ratios}

\item{conditions}{Character vector specifying conditions. By default it will assume names of `filter_cols` are the conditions prepended with "drop_".}
}
\value{
data frame with filtered data
}
\description{
Filter specified columns based on completeness for each condition.
}
\details{
This function filters the specified columns based on the completeness of the data.
The `filter_cols` argument specifies the columns of `dat` that contain missingness information (e.g. from `check_missingness()`).
These columns should be logical vectors with `TRUE` indicating too much missingness and that row should be filtered for the specified condition.
}
