% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptides.R
\name{process_peptides}
\alias{process_peptides}
\title{process_peptides
Process peptides from raw data}
\usage{
process_peptides(
  data,
  config,
  stage = file.path(config$output_dir, config$peptide_checkpoint),
  save_intermediate = TRUE
)
}
\arguments{
\item{data}{MSstats formatted data}

\item{config}{list of configuration parameters}

\item{stage}{character path to checkpoint file}

\item{save_intermediate}{logical save intermediate data}
}
\value{
data frame of processed peptide data
}
\description{
process_peptides
Process peptides from raw data
}
\details{
This function processes MSstats formatted data and returns peptides ready for ProtResDash. If save_intermediate is TRUE, the processed data are also saved to the checkpoint file.
}
