% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_raw}
\alias{process_raw}
\title{process_raw
Process raw data into MSstats format}
\usage{
process_raw(
  config = configure_formatR(),
  raw = NULL,
  save_intermediate = TRUE,
  ...
)
}
\arguments{
\item{config}{list of configuration parameters}

\item{raw}{data.frame of raw data. If not provided (default), `config$in_file` is loaded.}

\item{save_intermediate}{logical save intermediate data}

\item{...}{other parameters passed to `updt_config`}
}
\value{
data frame of processed raw data
}
\description{
process_raw
Process raw data into MSstats format
}
\details{
This function processes raw data into an MSstats formatted object in R.
If save_intermediate is TRUE, the processed data are also saved to the checkpoint file.
This also has the side effect of changing/updating `config` in the calling environment.

Any parameters that are `NULL` will be taken from the package defaults.
}
