% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_wb.R
\name{process_wb}
\alias{process_wb}
\title{process_wb
Format peptides and proteins for export to Excel}
\usage{
process_wb(
  proteins,
  peptides,
  config,
  stage = file.path(config$output_dir, config$wb_checkpoint),
  save_intermediate = TRUE,
  n_proteins = NULL
)
}
\arguments{
\item{proteins}{data frame of proteins data}

\item{peptides}{data frame of peptides data}

\item{config}{list of configuration parameters}

\item{stage}{character path to checkpoint file}

\item{save_intermediate}{logical save intermediate data}

\item{n_proteins}{integer number of proteins to process}
}
\value{
An excel workbook object
}
\description{
process_wb
Format peptides and proteins for export to Excel
}
\details{
This function processes formatted protein and peptide data and returns an excel workbook object ready for export to Excel. If save_intermediate is TRUE, the processed data are also saved to the checkpoint file.
}
